requirements.txt
selenium
pytest
python-dotenv
bs4